// src/interfaces/httpstatus.ts
export enum HttpStatusCode {
    OK = 200,
    NOT_SAVE =203,
    BAD_REQUEST = 400,
    VALIDATION_ERROR=422,
    NOT_FOUND = 404,
    INTERNAL_SERVER = 500,
    PAYMENT_REQUIRED = 402,
    NO_PERMISSION = 403,
    METHOD_NOT_ALLOWED = 405,
    NOT_ACCEPTABLE = 406,
    REQUEST_TIMEOUT = 408,
    UNAUTHORIZED = 401

}